+++ src/components/composites/CompositeItem.tsx
@@ -151,12 +151,31 @@ export const CompositeItem: React.FC<CompositeItemProps> = (props: CompositeItem
               onChange={() => setComposite({ ...composite, clickThroughSanitize: !composite.clickThroughSanitize })}
             />
           </Field>
-          <Field label="Open in New Tab" description="Open link in new tab" disabled={!composite.showComposite}>
+          <Field label="Open URL in New Tab" description="Open link in new tab" disabled={!composite.showComposite}>
             <Switch
               transparent={false}
               value={composite.clickThroughOpenNewTab}
               disabled={!composite.showComposite}
               onChange={() => setComposite({ ...composite, clickThroughOpenNewTab: !composite.clickThroughOpenNewTab })}
+            />
+          </Field>
+          <Field label="Enable Custom URL Target" description="Enable custom target" disabled={!composite.showComposite} hidden={composite.clickThroughOpenNewTab}>
+            <Switch
+              transparent={false}
+              value={composite.clickThroughCustomTargetEnabled}
+              disabled={!composite.showComposite}
+              onChange={() => setComposite({ ...composite,
+                clickThroughCustomTargetEnabled: !composite.clickThroughCustomTargetEnabled,
+                clickThroughCustomTarget: composite.clickThroughCustomTarget || ''
+              })}
+            />
+          </Field>
+          <Field label="Custom URL Target" description="Specify a custom target, typical values are: _blank|_self|_parent|_top|framename" disabled={!composite.showComposite} hidden={!composite.clickThroughCustomTargetEnabled}>
+            <Input
+              value={composite.clickThroughCustomTarget}
+              placeholder="_self"
+              disabled={!composite.showComposite}
+              onChange={(e) => setComposite({ ...composite, clickThroughCustomTarget: e.currentTarget.value })}
             />
           </Field>
           <Field disabled={!composite.showComposite}>