+++ src/module.ts
@@ -458,7 +458,7 @@ export const plugin = new PanelPlugin<PolystatOptions>(PolystatPanel)
       })
       // sanitize clickthrough
       .addBooleanSwitch({
-        name: 'Sanitize',
+        name: 'Sanitize URL',
         path: 'globalClickthroughSanitizedEnabled',
         defaultValue: true,
         category: ['Global'],
@@ -466,11 +466,28 @@ export const plugin = new PanelPlugin<PolystatOptions>(PolystatPanel)
       })
       // open in new tab
       .addBooleanSwitch({
-        name: 'Open In New Tab',
+        name: 'Open URL In New Tab',
         path: 'globalClickthroughNewTabEnabled',
         defaultValue: true,
         category: ['Global'],
         description: 'Opens clickthrough in a new tab',
+      })
+      .addBooleanSwitch({
+        name: 'Enable Custom URL Target',
+        path: 'globalClickthroughCustomTargetEnabled',
+        defaultValue: false,
+        category: ['Global'],
+        description: 'Use custom target for global clickthrough (this overrides the new tab setting above). Typical values are: _blank|_self|_parent|_top|framename',
+        showIf: (c) => c.globalClickthroughNewTabEnabled === false,
+      })
+      .addTextInput({
+        name: 'Custom URL Target',
+        path: 'globalClickthroughCustomTarget',
+        description:
+          'Provide a different target to be set in the target attribute of the clickthrough. Typical values are: _blank|_self|_parent|_top|framename',
+        category: ['Global'],
+        defaultValue: '',
+        showIf: (c) => c.globalClickthroughCustomTargetEnabled === true,
       })
 
       // aliasing